+++ .travis.yml
@@ -1,15 +1,13 @@
 language: go
 
 go:
-  - 1.7.x # See README.md for current minimum version.
-  - 1.8.x
-  - 1.9.x
+  - 1.9.x # See README.md for current minimum version.
   - 1.10.x
   - 1.11.x
   - 1.12.x
 
 script:
   - make check_license unused test-short
-  - if [[ ! $TRAVIS_GO_VERSION =~ ^1\.(7|8|9)\.[x0-9]+$ ]]; then make lint; fi
-  # style is only checked against the latest supported Go version.
+  - if [[ ! $TRAVIS_GO_VERSION =~ ^1\.9\.[x0-9]+$ ]]; then make lint; fi
+  # Style is only checked against the latest supported Go version.
   - if [[ $TRAVIS_GO_VERSION =~ ^1\.(12)\. ]]; then make style; fi
